Refactor getting flattened GQL input
This will make it easier for other input exceptions to use this info
